servicio-ocr/requirements.txt
Flask==3.0.0
google-cloud-vision==3.4.4
google-cloud-storage==2.14.0
google-cloud-pubsub==2.18.4
gunicorn==21.2.0
